https://github.com/mage2pro/core/issues/385

+++ ResponseValidator.php
@@ -6,11 +6,11 @@ final class ResponseValidator extends \Df\API\Response\Validator {
 	/**
 	 * 2015-10-10
 	 * @override
-	 * @see \Df\API\Response\Validator::long()
+	 * @see \Df\API\Exception::short()
 	 * @used-by \Df\API\Client::_p()
 	 * @used-by \Df\API\Exception::message()
 	 */
-	function long():string {
+	function short():string {
 		$e = $this->r(); /** @var array(string => string) $e */
 		return "Facebook API error of type {$e['type']}: «{$e['message']}».";
 	}
